internal: Centralize state initialization for entityMeta
