+++ packages/core/src/react-integration/__tests__/useResource-endpoint.web.tsx
@@ -8,6 +8,7 @@ import {
   ContextAuthdArticle,
   AuthContext,
 } from '__tests__/new';
+import { createEntityMeta } from '__tests__/utils';
 import { State } from '@rest-hooks/core';
 import { initialState } from '@rest-hooks/core/state/reducer';
 import React, { Suspense } from 'react';
@@ -254,18 +255,11 @@ describe('useResource()', () => {
   });
   it('should NOT suspend even when result is stale and options.invalidIfStale is false', () => {
     const { entities, result } = normalize(payload, CoolerArticleResource);
-    const entityMeta: any = {};
-    for (const k in entities) {
-      entityMeta[k] = {};
-      for (const pk in entities[k]) {
-        entityMeta[k][pk] = { date: 0, expiresAt: 0 };
-      }
-    }
     const fetchKey = CoolerArticleResource.detail().getFetchKey(payload);
     const state = {
       ...initialState,
       entities,
-      entityMeta,
+      entityMeta: createEntityMeta(entities),
       results: {
         [fetchKey]: result,
       },
@@ -300,20 +294,13 @@ describe('useResource()', () => {
     const fetchKey = InvalidIfStaleArticleResource.detail().getFetchKey(
       payload,
     );
-    const entityMeta: any = {};
-    for (const k in entities) {
-      entityMeta[k] = {};
-      for (const pk in entities[k]) {
-        entityMeta[k][pk] = { date: 0, expiresAt: 0 };
-      }
-    }
     const state = {
       ...initialState,
       entities,
       results: {
         [fetchKey]: result,
       },
-      entityMeta,
+      entityMeta: createEntityMeta(entities),
       meta: {
         [fetchKey]: {
           date: Infinity,
@@ -343,20 +330,13 @@ describe('useResource()', () => {
     const fetchKey = InvalidIfStaleArticleResource.detail().getFetchKey(
       payload,
     );
-    const entityMeta: any = {};
-    for (const k in entities) {
-      entityMeta[k] = {};
-      for (const pk in entities[k]) {
-        entityMeta[k][pk] = { date: 0, expiresAt: 0 };
-      }
-    }
     const state = {
       ...initialState,
       entities,
       results: {
         [fetchKey]: result,
       },
-      entityMeta,
+      entityMeta: createEntityMeta(entities),
       meta: {
         [fetchKey]: {
           date: 0,
