+++ packages/core/src/state/selectors/useDenormalized.ts
@@ -102,7 +102,7 @@ export default function useDenormalized<
       // using Object.keys ensures we don't hit `toString` type members
       Object.keys(resolvedEntities).forEach(key =>
         Object.keys(resolvedEntities[key]).forEach(pk => {
-          expiresAt = Math.min(expiresAt, entityMeta?.[key]?.[pk].expiresAt);
+          expiresAt = Math.min(expiresAt, entityMeta[key][pk].expiresAt);
         }),
       );
     } else {
